agora/plugins/requirements.txt
greenlet==0.4.17
joblib==0.17.0
msgpack==1.0.0
numpy==1.19.3
pandas==1.1.4
pyDOE2==1.3.0
pynvim==0.4.2
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.1
scikit-learn==0.23.2
scipy==1.5.3
six==1.15.0
threadpoolctl==2.1.0
